Issue #14436: Enforced naming convention
