+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/magicnumber/InputMagicNumberCheckDefault.java
@@ -1,8 +1,8 @@
 /*
 MagicNumber
 ignoreNumbers = (default)-1, 0, 1, 2
-ignoreHashCodeMethod = true
-ignoreAnnotation = true
+ignoreHashCodeMethod = (default)false
+ignoreAnnotation = (default)false
 ignoreFieldDeclaration = (default)false
 ignoreAnnotationElementDefaults = (default)true
 constantWaiverParentToken = (default)TYPECAST, METHOD_CALL, EXPR, ARRAY_INIT, UNARY_MINUS, \
@@ -15,11 +15,11 @@ tokens = (default)NUM_DOUBLE, NUM_FLOAT, NUM_INT, NUM_LONG
 package com.puppycrawl.tools.checkstyle.checks.coding.magicnumber;
 
 /**
- * Describe class InputMagicNumber
+ * Describe class InputMagicNumberDefaults
  * @author Rick Giles
  * @version 6-May-2003
  */
-public class InputMagicNumberIgnoreHashCodeMethod1 {
+public class InputMagicNumberCheckDefault {
     public void magicMethod() {
 
         final int INT_CONST = 101_000;
@@ -30,7 +30,6 @@ public class InputMagicNumberIgnoreHashCodeMethod1 {
         final double DOUBLE_CONST1 = 1.500_0D;
         final double DOUBLE_CONST2 = 1.5d;
         final double DOUBLE_CONST3 = 1.5;
-
 
         int int_var1 = 1;
         int int_var2 = (2);
@@ -58,7 +57,7 @@ public class InputMagicNumberIgnoreHashCodeMethod1 {
 
         int_array = new int[3]; // violation
 
-        int_magic1 += 3; // violation
+        int_magic1 += 3;  // violation
         double_magic1 *= 1.5; // violation
 
         for (int j = 3; j < 5; j += 3) { // 3 violations
@@ -69,31 +68,29 @@ public class InputMagicNumberIgnoreHashCodeMethod1 {
             int_magic1 = int_magic1 + 3; // violation
         }
 
-
         int octalVar0 = 00;
         int octalVar8 = 010; // violation
         int octalVar9 = 011; // violation
 
         long longOctalVar8 = 0_10L; // violation
-        long longOctalVar9 = 011l; // violation
-
+        long longOctalVar9 = 011l;  // violation
 
         int hexVar0 = 0x0;
         int hexVar16 = 0x10; // violation
-        int hexVar17 = 0X011;  // violation
+        int hexVar17 = 0X011; // violation
         long longHexVar0 = 0x0L;
-        long longHexVar16 = 0x10L;  // violation
+        long longHexVar16 = 0x10L; // violation
         long longHexVar17 = 0X11l; // violation
     }
 }
 
-interface Blah2IgnoreHashCodeMethod1
+interface Blah2CheckDefault
 {
   int LOW = 5;
   int HIGH = 78;
 }
 
-class ArrayMagicTestIgnoreHashCodeMethod1
+class ArrayMagicTestCheckDefault
 {
     private static final int[] NONMAGIC = {3};
     private int[] magic = {3}; // violation
