+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/magicnumber/InputMagicNumberCheckHashCodeValue.java
@@ -1,8 +1,8 @@
 /*
 MagicNumber
-ignoreNumbers =
+ignoreNumbers = (default)-1, 0, 1, 2
 ignoreHashCodeMethod = (default)false
-ignoreAnnotation = true
+ignoreAnnotation = (default)false
 ignoreFieldDeclaration = (default)false
 ignoreAnnotationElementDefaults = (default)true
 constantWaiverParentToken = (default)TYPECAST, METHOD_CALL, EXPR, ARRAY_INIT, UNARY_MINUS, \
@@ -14,7 +14,7 @@ tokens = (default)NUM_DOUBLE, NUM_FLOAT, NUM_INT, NUM_LONG
 
 package com.puppycrawl.tools.checkstyle.checks.coding.magicnumber;
 
-class InputMagicNumberIgnoreNone3 {
+public class InputMagicNumberCheckHashCodeValue {
 
     public int hashCode() {
         return 31; // violation
@@ -38,49 +38,49 @@ class InputMagicNumberIgnoreNone3 {
         int y=37; // violation
     }
 
-    public InputMagicNumberIgnoreNone3() {
+    public InputMagicNumberCheckHashCodeValue() {
         int z=101; // violation
     }
 
-    @InputMagicNumberIntMethodAnnotation(42)
+    @InputMagicNumberIntMethodAnnotation(42) // violation
     public void another() {
     }
 
-    @InputMagicNumberIntMethodAnnotation(value=43)
+    @InputMagicNumberIntMethodAnnotation(value=43) // violation
     public void another2() {
     }
 
-    @InputMagicNumberIntMethodAnnotation(-44)
+    @InputMagicNumberIntMethodAnnotation(-44) // violation
     public void anotherNegative() {
     }
 
-    @InputMagicNumberIntMethodAnnotation(value=-45)
+    @InputMagicNumberIntMethodAnnotation(value=-45) // violation
     public void anotherNegative2() {
     }
 }
 
-class TestMethodCallIgnoreNone3 {
+class TestMethodCallHashCodeValue {
 
-        public TestMethodCallIgnoreNone3(int x){
+        public TestMethodCallHashCodeValue(int x){
 
     }
 
         public void method2() {
-        final TestMethodCallIgnoreNone3 dummyObject = new TestMethodCallIgnoreNone3(62);
+        final TestMethodCallHashCodeValue dummyObject = new TestMethodCallHashCodeValue(62);
         }
 }
 
-class BinaryIgnoreNone3 {
+class BinaryCheckHashCodeValue {
     int intValue = 0b101; // violation
     long l = 0b1010000101000101101000010100010110100001010001011010000101000101L; // violation
 }
-@interface AnnotationWithDefaultValueIgnoreNone3 {
+@interface AnnotationWithDefaultValueCheckHashCodeValue {
     int value() default 101;
     int[] ar() default {102};
 }
-class AIgnoreNone3 {
+class ADefault3 {
     {
-        switch (Blah2IgnoreNone1.LOW) {
+        switch (Blah2CheckDefault.LOW) {
         default:
             int b = 122; // violation
         }
