+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/magicnumber/InputMagicNumberCheckIgnoreFieldDeclarationBinaryAnnotation.java
@@ -1,0 +1,89 @@
+/*
+MagicNumber
+ignoreNumbers = (default)-1, 0, 1, 2
+ignoreHashCodeMethod = (default)false
+ignoreAnnotation = (default)false
+ignoreFieldDeclaration = true
+ignoreAnnotationElementDefaults = (default)true
+constantWaiverParentToken = (default)TYPECAST, METHOD_CALL, EXPR, ARRAY_INIT, UNARY_MINUS, \
+                            UNARY_PLUS, ELIST, STAR, ASSIGN, PLUS, MINUS, DIV, LITERAL_NEW
+tokens = (default)NUM_DOUBLE, NUM_FLOAT, NUM_INT, NUM_LONG
+
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.coding.magicnumber;
+
+class InputMagicNumberCheckIgnoreFieldDeclarationBinaryAnnotation {
+
+    public int hashCode() {
+        return 31; // violation
+    }
+
+
+    public int hashCode(int val) {
+        return 42; // violation
+    }
+
+
+    public int hashcode() {
+        return 13; // violation
+    }
+
+    static {
+        int x=21; // violation
+    }
+
+    {
+        int y=37; // violation
+    }
+
+    public InputMagicNumberCheckIgnoreFieldDeclarationBinaryAnnotation() {
+        int z=101; // violation
+    }
+
+    @InputMagicNumberIntMethodAnnotation(42) // violation
+    public void another() {
+    }
+
+    @InputMagicNumberIntMethodAnnotation(value=43) // violation
+    public void another2() {
+    }
+
+    @InputMagicNumberIntMethodAnnotation(-44) // violation
+    public void anotherNegative() {
+    }
+
+    @InputMagicNumberIntMethodAnnotation(value=-45) // violation
+    public void anotherNegative2() {
+    }
+}
+
+class TestMethodCallIgnoreFieldDeclarationBinaryAnnotation {
+
+        public TestMethodCallIgnoreFieldDeclarationBinaryAnnotation(int x){
+
+    }
+
+        public void method2() {
+        final TestMethodCallIgnoreFieldDeclarationBinaryAnnotation dummyObject =
+                new TestMethodCallIgnoreFieldDeclarationBinaryAnnotation(62);
+        }
+}
+
+class BinaryIgnoreFieldDeclarationBinaryAnnotation {
+    int intValue = 0b101;
+    long l = 0b1010000101000101101000010100010110100001010001011010000101000101L;
+}
+@interface AnnotationWithDefaultValueIgnoreFieldDeclarationBinaryAnnotation {
+    int value() default 101;
+    int[] ar() default {102};
+}
+class AIgnoreFieldDeclarationBinaryAnnotation {
+    {
+        switch (Blah2IgnoreFieldDeclarationViolationHandler.LOW) {
+        default:
+            int b = 122; // violation
+        }
+    }
+}
