+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/magicnumber/InputMagicNumberCheckIgnoreHashCodeMethod3.java
@@ -1,9 +1,9 @@
 /*
 MagicNumber
 ignoreNumbers = (default)-1, 0, 1, 2
-ignoreHashCodeMethod = (default)false
-ignoreAnnotation = (default)false
-ignoreFieldDeclaration = true
+ignoreHashCodeMethod = true
+ignoreAnnotation = true
+ignoreFieldDeclaration = (default)false
 ignoreAnnotationElementDefaults = (default)true
 constantWaiverParentToken = (default)TYPECAST, METHOD_CALL, EXPR, ARRAY_INIT, UNARY_MINUS, \
                             UNARY_PLUS, ELIST, STAR, ASSIGN, PLUS, MINUS, DIV, LITERAL_NEW
@@ -14,10 +14,10 @@ tokens = (default)NUM_DOUBLE, NUM_FLOAT, NUM_INT, NUM_LONG
 
 package com.puppycrawl.tools.checkstyle.checks.coding.magicnumber;
 
-class InputMagicNumberIgnoreFieldDeclaration3 {
+class InputMagicNumberCheckIgnoreHashCodeMethod3 {
 
     public int hashCode() {
-        return 31; // violation
+        return 31;
     }
 
 
@@ -38,50 +38,50 @@ class InputMagicNumberIgnoreFieldDeclaration3 {
         int y=37; // violation
     }
 
-    public InputMagicNumberIgnoreFieldDeclaration3() {
+    public InputMagicNumberCheckIgnoreHashCodeMethod3() {
         int z=101; // violation
     }
 
-    @InputMagicNumberIntMethodAnnotation(42) // violation
+    @InputMagicNumberIntMethodAnnotation(42)
     public void another() {
     }
 
-    @InputMagicNumberIntMethodAnnotation(value=43) // violation
+    @InputMagicNumberIntMethodAnnotation(value=43)
     public void another2() {
     }
 
-    @InputMagicNumberIntMethodAnnotation(-44) // violation
+    @InputMagicNumberIntMethodAnnotation(-44)
     public void anotherNegative() {
     }
 
-    @InputMagicNumberIntMethodAnnotation(value=-45) // violation
+    @InputMagicNumberIntMethodAnnotation(value=-45)
     public void anotherNegative2() {
     }
 }
 
-class TestMethodCallIgnoreFieldDeclaration3 {
+class TestMethodCallIgnoreHashCodeMethod3 {
 
-        public TestMethodCallIgnoreFieldDeclaration3(int x){
+        public TestMethodCallIgnoreHashCodeMethod3(int x){
 
     }
 
         public void method2() {
-        final TestMethodCallIgnoreFieldDeclaration3 dummyObject =
-                new TestMethodCallIgnoreFieldDeclaration3(62);
+        final TestMethodCallIgnoreHashCodeMethod3 dummyObject =
+                new TestMethodCallIgnoreHashCodeMethod3(62);
         }
 }
 
-class BinaryIgnoreFieldDeclaration3 {
-    int intValue = 0b101;
-    long l = 0b1010000101000101101000010100010110100001010001011010000101000101L;
+class BinaryIgnoreHashCodeMethod3 {
+    int intValue = 0b101; // violation
+    long l = 0b1010000101000101101000010100010110100001010001011010000101000101L; // violation
 }
-@interface AnnotationWithDefaultValueIgnoreFieldDeclaration3 {
+@interface AnnotationWithDefaultValueIgnoreHashCodeMethod3{
     int value() default 101;
     int[] ar() default {102};
 }
-class AIgnoreFieldDeclaration3 {
+class AIgnoreHashCodeMethod3 {
     {
-        switch (Blah2IgnoreFieldDeclaration1.LOW) {
+        switch (Blah2IgnoreHashCodeMethod1.LOW) {
         default:
             int b = 122; // violation
         }
