+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/magicnumber/InputMagicNumberCheckIgnoreNegativeOctalHex1.java
@@ -1,13 +1,13 @@
 /*
 MagicNumber
-ignoreNumbers = 0, 1, 3.0, 8, 16, 3000
+ignoreNumbers = -9223372036854775808, -2147483648, -1, 0, 1, 2, -2
 ignoreHashCodeMethod = (default)false
 ignoreAnnotation = true
 ignoreFieldDeclaration = (default)false
 ignoreAnnotationElementDefaults = (default)true
 constantWaiverParentToken = (default)TYPECAST, METHOD_CALL, EXPR, ARRAY_INIT, UNARY_MINUS, \
                             UNARY_PLUS, ELIST, STAR, ASSIGN, PLUS, MINUS, DIV, LITERAL_NEW
-tokens = (default)NUM_DOUBLE, NUM_FLOAT, NUM_INT, NUM_LONG
+tokens = NUM_INT, NUM_LONG
 
 
 */
@@ -19,9 +19,9 @@ package com.puppycrawl.tools.checkstyle.checks.coding.magicnumber;
  * @author Rick Giles
  * @version 6-May-2003
  */
-public class InputMagicNumberIgnoreSome1 {
+public class InputMagicNumberCheckIgnoreNegativeOctalHex1 {
     public void magicMethod() {
-        //constants, ignore
+
         final int INT_CONST = 101_000;
         final long LONG_CONST1 = 100_000L;
         final long LONG_CONST2 = 100l;
@@ -31,71 +31,71 @@ public class InputMagicNumberIgnoreSome1 {
         final double DOUBLE_CONST2 = 1.5d;
         final double DOUBLE_CONST3 = 1.5;
 
-        //ignore by default
+
         int int_var1 = 1;
-        int int_var2 = (2); // violation
+        int int_var2 = (2);
         long long_var1 = 0L;
         long long_var2 = 0l;
         double double_var1 = 0D;
         double double_var2 = 0d;
 
-        int[] int_array = new int[2]; // violation
+        int[] int_array = new int[2];
 
-        int_var1 = 1 + 2; // violation
+        int_var1 = 1 + 2;
         int_var1 += 1;
-        double_var1 = 1.0 + 2.0; // violation
+        double_var1 = 1.0 + 2.0;
 
-        for (int i = 0; i < 2; i++); // violation
+        for (int i = 0; i < 2; i++);
 
-        if (1 < 2); // violation
+        if (1 < 2);
 
-        if (1.0 < 2.0); // violation
+        if (1.0 < 2.0);
 
-        //magic numbers
-        int int_magic1 = 3_000;
-        double double_magic1 = 1.5_0; // violation
-        int int_magic2 = (3 + 4); // violation
 
-        int_array = new int[3];
+        int int_magic1 = 3_000; // violation
+        double double_magic1 = 1.5_0;
+        int int_magic2 = (3 + 4); // 2 violations
 
-        int_magic1 += 3;
-        double_magic1 *= 1.5; // violation
+        int_array = new int[3]; // violation
 
-        for (int j = 3; j < 5; j += 3) { // violation
+        int_magic1 += 3; // violation
+        double_magic1 *= 1.5;
+
+        for (int j = 3; j < 5; j += 3) { // 3 violations
             int_magic1++;
         }
 
-        if (int_magic1 < 3) {
-            int_magic1 = int_magic1 + 3;
+        if (int_magic1 < 3) { // violation
+            int_magic1 = int_magic1 + 3; // violation
         }
 
 
         int octalVar0 = 00;
-        int octalVar8 = 010;
+        int octalVar8 = 010; // violation
         int octalVar9 = 011; // violation
 
-        long longOctalVar8 = 0_10L;
+        long longOctalVar8 = 0_10L; // violation
         long longOctalVar9 = 011l; // violation
 
 
         int hexVar0 = 0x0;
-        int hexVar16 = 0x10;
+        int hexVar16 = 0x10; // violation
         int hexVar17 = 0X011;  // violation
         long longHexVar0 = 0x0L;
-        long longHexVar16 = 0x10L;
+        long longHexVar16 = 0x10L;  // violation
         long longHexVar17 = 0X11l; // violation
     }
 }
 
-interface Blah2IgnoreSome1
+interface Blah2IgnoreNegativeOctalHex1
 {
   int LOW = 5;
   int HIGH = 78;
 }
 
-class ArrayMagicTestIgnoreSome1
+class ArrayMagicTestIgnoreNegativeOctalHex1
 {
     private static final int[] NONMAGIC = {3};
-    private int[] magic = {3};
+    private int[] magic = {3}; // violation
     private static final int[][] NONMAGIC2 = {{1}, {2}, {3}};
 }
