+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/magicnumber/InputMagicNumberCheckIntegersOnlyBinaryAndAnnotation.java
@@ -2,19 +2,19 @@
 MagicNumber
 ignoreNumbers = (default)-1, 0, 1, 2
 ignoreHashCodeMethod = (default)false
-ignoreAnnotation = (default)false
+ignoreAnnotation = true
 ignoreFieldDeclaration = (default)false
 ignoreAnnotationElementDefaults = (default)true
-constantWaiverParentToken = ASSIGN, ARRAY_INIT, EXPR, UNARY_PLUS, UNARY_MINUS, TYPECAST, \
-                            ELIST, STAR, DIV, PLUS, MINUS
-tokens = (default)NUM_DOUBLE, NUM_FLOAT, NUM_INT, NUM_LONG
+constantWaiverParentToken = (default)TYPECAST, METHOD_CALL, EXPR, ARRAY_INIT, UNARY_MINUS, \
+                            UNARY_PLUS, ELIST, STAR, ASSIGN, PLUS, MINUS, DIV, LITERAL_NEW
+tokens = NUM_INT, NUM_LONG
 
 
 */
 
 package com.puppycrawl.tools.checkstyle.checks.coding.magicnumber;
 
-class InputMagicNumberWaiverParentToken3 {
+class InputMagicNumberCheckIntegersOnlyBinaryAndAnnotation {
 
     public int hashCode() {
         return 31; // violation
@@ -38,55 +38,51 @@ class InputMagicNumberWaiverParentToken3 {
         int y=37; // violation
     }
 
-    public InputMagicNumberWaiverParentToken3() {
+    public InputMagicNumberCheckIntegersOnlyBinaryAndAnnotation() {
         int z=101; // violation
     }
 
-    @InputMagicNumberIntMethodAnnotation(42) // violation
+    @InputMagicNumberIntMethodAnnotation(42)
     public void another() {
     }
 
-    @InputMagicNumberIntMethodAnnotation(value=43) // violation
+    @InputMagicNumberIntMethodAnnotation(value=43)
     public void another2() {
     }
 
-    @InputMagicNumberIntMethodAnnotation(-44) // violation
+    @InputMagicNumberIntMethodAnnotation(-44)
     public void anotherNegative() {
     }
 
-    @InputMagicNumberIntMethodAnnotation(value=-45)  // violation
+    @InputMagicNumberIntMethodAnnotation(value=-45)
     public void anotherNegative2() {
     }
 }
 
-class TestMethodCallWaiverParentToken3 {
+class TestMethodBinaryAndAnnotation {
 
-        public TestMethodCallWaiverParentToken3(int x){
+        public TestMethodBinaryAndAnnotation(int x){
 
     }
 
         public void method2() {
-        final TestMethodCallWaiverParentToken3 dummyObject =
-                new TestMethodCallWaiverParentToken3(62); // violation
+        final TestMethodBinaryAndAnnotation dummyObject = new TestMethodBinaryAndAnnotation(62);
         }
 }
 
-class BinaryWaiverParentToken3 {
+class BinaryIntegersOnlyBinaryAndAnnotation {
     int intValue = 0b101; // violation
     long l = 0b1010000101000101101000010100010110100001010001011010000101000101L; // violation
 }
-@interface AnnotationWithDefaultValueWaiverParentToken3 {
+@interface AnnotationWithDefaultValueIntegersOnlyBinaryAndAnnotation {
     int value() default 101;
     int[] ar() default {102};
 }
-class A {
+class AIntegersOnlyBinaryAndAnnotation {
     {
-        switch (Blah2WaiverParentToken1.LOW) {
+        switch (Blah2IntegersOnlyViolationHandler.LOW) {
         default:
             int b = 122; // violation
         }
     }
-}
-@interface InputMagicNumberIntMethodAnnotation {
-        int value();
 }
