+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/magicnumber/InputMagicNumberCheckIntegersOnlyHexSignedAndOctal.java
@@ -1,20 +1,19 @@
 /*
 MagicNumber
-ignoreNumbers = 0, 1, 3.0, 8, 16, 3000
+ignoreNumbers = (default)-1, 0, 1, 2
 ignoreHashCodeMethod = (default)false
 ignoreAnnotation = true
 ignoreFieldDeclaration = (default)false
 ignoreAnnotationElementDefaults = (default)true
 constantWaiverParentToken = (default)TYPECAST, METHOD_CALL, EXPR, ARRAY_INIT, UNARY_MINUS, \
                             UNARY_PLUS, ELIST, STAR, ASSIGN, PLUS, MINUS, DIV, LITERAL_NEW
-tokens = (default)NUM_DOUBLE, NUM_FLOAT, NUM_INT, NUM_LONG
+tokens = NUM_INT, NUM_LONG
 
 
 */
 
 package com.puppycrawl.tools.checkstyle.checks.coding.magicnumber;
-
-public class InputMagicNumberIgnoreSome2 {
+public class InputMagicNumberCheckIntegersOnlyHexSignedAndOctal {
     /**
      * test long hex
      */
@@ -26,23 +25,24 @@ public class InputMagicNumberIgnoreSome2 {
      */
     public static final int CONST_PLUS_THREE = +3;
     public static final int CONST_MINUS_TWO = -2;
-    private int mPlusThree = +3;
+    private int mPlusThree = +3; // violation
     private int mMinusTwo = -2; // violation
-    private double mPlusDecimal = +3.5; // violation
-    private double mMinusDecimal = -2.5; // violation
+    private double mPlusDecimal = +3.5;
+    private double mMinusDecimal = -2.5;
 
     /**
      * test octal and hex negative values
      */
-    private int hexIntMinusOne = 0xffffffff; // violation
-    private long hexLongMinusOne = 0xffffffffffffffffL; // violation
+    private int hexIntMinusOne = 0xffffffff;
+    private long hexLongMinusOne = 0xffffffffffffffffL;
     private long hexIntMinValue = 0x80000000; // violation
     private long hexLongMinValue = 0x8000000000000000L; // violation
-    private int octalIntMinusOne = 037777777777;  // violation
-    private long octalLongMinusOne = 01777777777777777777777L;  // violation
+    private int octalIntMinusOne = 037777777777;
+    private long octalLongMinusOne = 01777777777777777777777L;
     private long octalIntMinValue = 020000000000;  // violation
     private long octalLongMinValue = 01000000000000000000000L; // violation
     public static final int TESTINTVAL = (byte) 0x80;
+
     public static final java.util.List MYLIST = new java.util.ArrayList() {
         public int size() {
 
@@ -56,11 +56,11 @@ public class InputMagicNumberIgnoreSome2 {
     public final javax.swing.border.Border StdBorder =
             javax.swing.BorderFactory.createEmptyBorder(3, 3, 3, 3);
 
-    enum MyEnum2IgnoreSome2 {
+    enum MyEnum2IntegersOnlyHexSignedAndOctal {
         A_2(3),
         B_2(54);
 
-        private MyEnum2IgnoreSome2(int value) {
+        private MyEnum2IntegersOnlyHexSignedAndOctal(int value) {
 
         }
     }
