+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/magicnumber/InputMagicNumberCheckLocalVariablesSecondary.java
@@ -17,13 +17,12 @@ com.puppycrawl.tools.checkstyle.checks.SuppressWarningsHolder
 
 package com.puppycrawl.tools.checkstyle.checks.coding.magicnumber;
 
-public class InputMagicNumberLocalVariables {
+public class InputMagicNumberCheckLocalVariablesSecondary {
     @SuppressWarnings("MagicNumber")
     private int A = 8;
 
-    @SuppressWarnings("MagicNumber")
     void method() {
-        int b = 8;
+        int b = 8; // violation
     }
 
     void method2() {
