+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/magicnumber/InputMagicNumberCheckMagicNumberNumericalDataHandler.java
@@ -19,7 +19,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-public class InputMagicNumberMagicNumber3 {
+public class InputMagicNumberCheckMagicNumberNumericalDataHandler {
 
     Object obj = new Object() {
