+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/magicnumber/InputMagicNumberCheckWaiverParentTokenViolationHandler.java
@@ -5,8 +5,8 @@ ignoreHashCodeMethod = (default)false
 ignoreAnnotation = (default)false
 ignoreFieldDeclaration = (default)false
 ignoreAnnotationElementDefaults = (default)true
-constantWaiverParentToken = (default)TYPECAST, METHOD_CALL, EXPR, ARRAY_INIT, UNARY_MINUS, \
-                            UNARY_PLUS, ELIST, STAR, ASSIGN, PLUS, MINUS, DIV, LITERAL_NEW
+constantWaiverParentToken = ASSIGN, ARRAY_INIT, EXPR, UNARY_PLUS, UNARY_MINUS, TYPECAST, \
+                            ELIST, STAR, DIV, PLUS, MINUS
 tokens = (default)NUM_DOUBLE, NUM_FLOAT, NUM_INT, NUM_LONG
 
 
@@ -15,11 +15,11 @@ tokens = (default)NUM_DOUBLE, NUM_FLOAT, NUM_INT, NUM_LONG
 package com.puppycrawl.tools.checkstyle.checks.coding.magicnumber;
 
 /**
- * Describe class InputMagicNumberDefaults
+ * Describe class InputMagicNumber
  * @author Rick Giles
  * @version 6-May-2003
  */
-public class InputMagicNumberDefault1 {
+public class InputMagicNumberCheckWaiverParentTokenViolationHandler {
     public void magicMethod() {
 
         final int INT_CONST = 101_000;
@@ -30,6 +30,7 @@ public class InputMagicNumberDefault1 {
         final double DOUBLE_CONST1 = 1.500_0D;
         final double DOUBLE_CONST2 = 1.5d;
         final double DOUBLE_CONST3 = 1.5;
+
 
         int int_var1 = 1;
         int int_var2 = (2);
@@ -53,14 +54,14 @@ public class InputMagicNumberDefault1 {
 
         int int_magic1 = 3_000; // violation
         double double_magic1 = 1.5_0; // violation
-        int int_magic2 = (3 + 4); // 2 violations
+        int int_magic2 = (3 + 4);  // 2 violations
 
-        int_array = new int[3]; // violation
+        int_array = new int[3];  // violation
 
-        int_magic1 += 3;  // violation
+        int_magic1 += 3;   // violation
         double_magic1 *= 1.5; // violation
 
-        for (int j = 3; j < 5; j += 3) { // 3 violations
+        for (int j = 3; j < 5; j += 3) {  // 3 violations
             int_magic1++;
         }
 
@@ -68,15 +69,17 @@ public class InputMagicNumberDefault1 {
             int_magic1 = int_magic1 + 3; // violation
         }
 
+
         int octalVar0 = 00;
-        int octalVar8 = 010; // violation
-        int octalVar9 = 011; // violation
+        int octalVar8 = 010;  // violation
+        int octalVar9 = 011;  // violation
 
         long longOctalVar8 = 0_10L; // violation
         long longOctalVar9 = 011l;  // violation
 
+
         int hexVar0 = 0x0;
-        int hexVar16 = 0x10; // violation
+        int hexVar16 = 0x10;  // violation
         int hexVar17 = 0X011; // violation
         long longHexVar0 = 0x0L;
         long longHexVar16 = 0x10L; // violation
@@ -84,15 +87,15 @@ public class InputMagicNumberDefault1 {
     }
 }
 
-interface Blah2Default1
+interface Blah2WaiverParentTokenViolationHandler
 {
   int LOW = 5;
   int HIGH = 78;
 }
 
-class ArrayMagicTestDefault1
+class ArrayMagicTestWaiverParentTokenViolationHandler
 {
     private static final int[] NONMAGIC = {3};
-    private int[] magic = {3}; // violation
+    private int[] magic = {3};  // violation
     private static final int[][] NONMAGIC2 = {{1}, {2}, {3}};
 }
