+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/magicnumber/InputMagicNumberIgnoreCheckHexSignedAndOctalValue.java
@@ -1,19 +1,20 @@
 /*
 MagicNumber
-ignoreNumbers = (default)-1, 0, 1, 2
+ignoreNumbers = 0, 1, 3.0, 8, 16, 3000
 ignoreHashCodeMethod = (default)false
-ignoreAnnotation = (default)false
+ignoreAnnotation = true
 ignoreFieldDeclaration = (default)false
 ignoreAnnotationElementDefaults = (default)true
-constantWaiverParentToken = ASSIGN, ARRAY_INIT, EXPR, UNARY_PLUS, UNARY_MINUS, TYPECAST, \
-                            ELIST, STAR, DIV, PLUS, MINUS
+constantWaiverParentToken = (default)TYPECAST, METHOD_CALL, EXPR, ARRAY_INIT, UNARY_MINUS, \
+                            UNARY_PLUS, ELIST, STAR, ASSIGN, PLUS, MINUS, DIV, LITERAL_NEW
 tokens = (default)NUM_DOUBLE, NUM_FLOAT, NUM_INT, NUM_LONG
 
 
 */
 
 package com.puppycrawl.tools.checkstyle.checks.coding.magicnumber;
-public class InputMagicNumberWaiverParentToken2 {
+
+public class InputMagicNumberIgnoreCheckHexSignedAndOctalValue {
     /**
      * test long hex
      */
@@ -25,24 +26,23 @@ public class InputMagicNumberWaiverParentToken2 {
      */
     public static final int CONST_PLUS_THREE = +3;
     public static final int CONST_MINUS_TWO = -2;
-    private int mPlusThree = +3; // violation
-    private int mMinusTwo = -2;  // violation
+    private int mPlusThree = +3;
+    private int mMinusTwo = -2; // violation
     private double mPlusDecimal = +3.5; // violation
     private double mMinusDecimal = -2.5; // violation
 
     /**
      * test octal and hex negative values
      */
-    private int hexIntMinusOne = 0xffffffff;
-    private long hexLongMinusOne = 0xffffffffffffffffL;
+    private int hexIntMinusOne = 0xffffffff; // violation
+    private long hexLongMinusOne = 0xffffffffffffffffL; // violation
     private long hexIntMinValue = 0x80000000; // violation
     private long hexLongMinValue = 0x8000000000000000L; // violation
-    private int octalIntMinusOne = 037777777777;
-    private long octalLongMinusOne = 01777777777777777777777L;
-    private long octalIntMinValue = 020000000000; // violation
-    private long octalLongMinValue = 01000000000000000000000L;  // violation
+    private int octalIntMinusOne = 037777777777;  // violation
+    private long octalLongMinusOne = 01777777777777777777777L;  // violation
+    private long octalIntMinValue = 020000000000;  // violation
+    private long octalLongMinValue = 01000000000000000000000L; // violation
     public static final int TESTINTVAL = (byte) 0x80;
-
     public static final java.util.List MYLIST = new java.util.ArrayList() {
         public int size() {
 
@@ -51,16 +51,16 @@ public class InputMagicNumberWaiverParentToken2 {
     };
 
     public final double SpecialSum = 2 + 1e10, SpecialDifference = 4 - java.lang.Math.PI;
-    public final Integer DefaultInit = new Integer(27); // violation
+    public final Integer DefaultInit = new Integer(27);
     public final int SpecsPerDay = 24 * 60 * 60, SpecialRatio = 4 / 3;
     public final javax.swing.border.Border StdBorder =
-            javax.swing.BorderFactory.createEmptyBorder(3, 3, 3, 3); // 4 violations
+            javax.swing.BorderFactory.createEmptyBorder(3, 3, 3, 3);
 
-    enum MyEnum2WaiverParentToken2 {
-        A(3),
-        B(54);
+    enum MyEnum2IgnoreCheckHexSignedAndOctalValue {
+        A_2(3),
+        B_2(54);
 
-        private MyEnum2WaiverParentToken2(int value) {
+        private MyEnum2IgnoreCheckHexSignedAndOctalValue(int value) {
 
         }
     }
